Moved the notification permission popup to user and organizer activities
