+++ app/src/main/java/com/example/mini_pekkas/OrganizerActivity.java
@@ -1,21 +1,27 @@
 package com.example.mini_pekkas;
 
+import android.Manifest;
+import android.content.pm.PackageManager;
+import android.os.Build;
 import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
-import android.widget.EditText;
+import android.widget.Toast;
 
+import androidx.activity.result.ActivityResultLauncher;
+import androidx.activity.result.contract.ActivityResultContracts;
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.content.ContextCompat;
 import androidx.fragment.app.Fragment;
 import androidx.navigation.NavController;
 import androidx.navigation.Navigation;
 import androidx.navigation.fragment.NavHostFragment;
 import androidx.navigation.ui.AppBarConfiguration;
 import androidx.navigation.ui.NavigationUI;
+import androidx.work.OneTimeWorkRequest;
+import androidx.work.WorkManager;
 
 import com.example.mini_pekkas.databinding.OrganizerMainBinding;
+import com.example.mini_pekkas.notification.NotificationWorker;
 import com.example.mini_pekkas.ui.event.organizer.EventChooseUsersFragment;
-import com.example.mini_pekkas.ui.event.organizer.EventDetailsFragment;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 /**
  * OrganizerActivity is the main activity for organizers, setting up the navigation controller
@@ -51,6 +57,8 @@ public class OrganizerActivity extends AppCompatActivity {
         // Set up the action bar and bottom navigation
         NavigationUI.setupActionBarWithNavController(OrganizerActivity.this, navController, appBarConfiguration);
         NavigationUI.setupWithNavController(binding.organizerView, navController);
+
+        askNotificationPermission();    // Ask for notification permissions
     }
 
 
@@ -65,5 +73,50 @@ public class OrganizerActivity extends AppCompatActivity {
             return true;
         }
         return navController.navigateUp();
+    }
+
+    /**
+     * Registers a callback to handle the result of the permission request. If the permission is granted, the notification listener is initialized.
+     */
+    private final ActivityResultLauncher<String> requestPermissionLauncher =
+            registerForActivityResult(new ActivityResultContracts.RequestPermission(), isGranted -> {
+                if (isGranted) {
+                    // FCM SDK (and your app) can post notifications.
+                    // Initialize the notification listener
+                    OneTimeWorkRequest notificationWorkRequest =
+                            new OneTimeWorkRequest.Builder(NotificationWorker.class)
+                                    .build();
+                    WorkManager.getInstance(this).enqueue(notificationWorkRequest);
+                } else {
+                    // Inform user that that your app will not show notifications. Make a toast
+                    Toast.makeText(this, "Notification permission denied", Toast.LENGTH_SHORT).show();
+                }
+            });
+
+    /**
+     * Asks for notification permission. If the permission is already granted, the notification listener is initialized.
+     */
+    private void askNotificationPermission() {
+        // This is only necessary for API level >= 33 (TIRAMISU)
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+            if (ContextCompat.checkSelfPermission(this, android.Manifest.permission.POST_NOTIFICATIONS) ==
+                    PackageManager.PERMISSION_GRANTED) {
+                // FCM SDK (and your app) can post notifications.
+                // Initialize the notification listener
+                OneTimeWorkRequest notificationWorkRequest =
+                        new OneTimeWorkRequest.Builder(NotificationWorker.class)
+                                .build();
+                WorkManager.getInstance(this).enqueue(notificationWorkRequest);
+            } else if (shouldShowRequestPermissionRationale(android.Manifest.permission.POST_NOTIFICATIONS)) {
+                // TODO: display an educational UI explaining to the user the features that will be enabled
+                //       by them granting the POST_NOTIFICATION permission. This UI should provide the user
+                //       "OK" and "No thanks" buttons. If the user selects "OK," directly request the permission.
+                //       If the user selects "No thanks," allow the user to continue without notifications.
+                //       This should be done if time permits
+            } else {
+                // Directly ask for the permission
+                requestPermissionLauncher.launch(Manifest.permission.POST_NOTIFICATIONS);
+            }
+        }
     }
 }
